#125 : Change formula compiler and heal/damage resolver to use decima…

+++ MysteryMud.Domain/Combat/Calculators/HealCalculator.cs
@@ -4,7 +4,7 @@ namespace MysteryMud.Domain.Combat.Calculators;
 
 public static class HealCalculator
 {
-    public static int ModifyHeal(Entity target, int healAmount, Entity source)
+    public static decimal ModifyHeal(Entity target, decimal healAmount, Entity source)
     {
         return healAmount; // TODO: apply any healing modifiers here (buffs, debuffs, etc.)
     }
