+++ MysteryMud.Domain/Combat/Effect/Definitions/PeriodicHealActionDefinition.cs
@@ -2,5 +2,5 @@
 
 public class PeriodicHealActionDefinition : EffectActionDefinition
 {
-    public required Func<EffectContext, int> AmountFunc { get; init; }
+    public required Func<EffectContext, decimal> AmountFunc { get; init; }
 }
